+++ lib/src/presentation/widgets/questionnaire_item/choice/questionnaire_check_box_choice_item_view.dart
@@ -26,7 +26,7 @@ class QuestionnaireCheckBoxChoiceItemViewState
           .map((entry) => CheckboxListTile(
                 controlAffinity: ListTileControlAffinity.leading,
                 contentPadding: EdgeInsets.zero,
-                title: Text(entry.title ?? ''),
+                title: Text(valueNameResolver(entry)),
                 value: isSelected(entry),
                 onChanged: (selected) =>
                     onSelectedValuesChanged(selected, entry),