+++ lib/src/presentation/widgets/questionnaire_item/choice/questionnaire_drop_down_choice_item_view.dart
@@ -28,7 +28,7 @@ class QuestionnaireDropDownChoiceItemViewState
       controller: controller,
       values: values,
       onChanged: onSelectedValueChanged,
-      nameResolver: (answer) => answer.title ?? '',
+      nameResolver: valueNameResolver,
     );
   }
 }