+++ lib/src/presentation/widgets/questionnaire_item/choice/questionnaire_radio_button_choice_item_view.dart
@@ -27,7 +27,7 @@ class QuestionnaireRadioButtonChoiceItemViewState
           .map((entry) => RadioListTile<QuestionnaireAnswerOption>(
                 controlAffinity: ListTileControlAffinity.leading,
                 contentPadding: EdgeInsets.zero,
-                title: Text(entry.title ?? ''),
+                title: Text(valueNameResolver(entry)),
                 value: entry,
                 groupValue: selectedValue,
                 onChanged: onSelectedValueChanged,