+++ CHANGELOG.md
@@ -9,6 +9,14 @@ Types of changes
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## 5.0.0
+### Added
+- Added support for **Translation** extension according to [FHIR Extension Translation](http://hl7.org/fhir/StructureDefinition/translation) definition. Thanks [sh1l0n](https://github.com/sh1l0n) [PR-18](https://github.com/luis901101/fhir_questionnaire/pull/18)
+- Added translation extension example to sampleGeneric Questionnaire. 
+
+### Changed
+- **Breaking Change**: changed `String? locale` to `Locale? locale` in `QuestionnaireView` and `QuestionnaireBaseLocalization` to allow better localization support.
+
 ## 4.0.0
 ### Changed
 - **Breaking Change**: `onGenerateItemResponse` callback changed to also provide the generated `QuestionnaireResponseItem` as an argument of this callback. This way when this callback is being implemented it is also possible to just modify the `QuestionnaireResponseItem` and return it. Thanks [easazade](https://github.com/easazade) [PR-12](https://github.com/luis901101/fhir_questionnaire/pull/12)